Extract shared prompt helpers into internal/agent/prompt (#619)
Move the reusable planning/review prompt templates, prompt rendering,
repository context gathering, output heuristics, and parse/validate
helpers out of internal/agent/claude into a new shared package
internal/agent/prompt so a second provider (OpenCode, future task) can
reuse them instead of duplicating ~250 lines.

The claude package now calls the exported helpers
(prompt.RenderPlanningPrompt, prompt.ReadPlanFile, prompt.ValidatePlan,
prompt.RenderReviewPrompt, prompt.ParseReviewOutput,
prompt.IsSuspiciousOutput, prompt.RetryDelay, prompt.WriteSystemPromptFile,
prompt.ReviewSystemPrompt). Subprocess orchestration (argv, exec.Cmd,
retry loops) stays in the claude package.

This is a behavior-preserving refactor: argv for Plan, Execute, Review,
and Discuss is unchanged; all existing tests still pass.
