+++ .herd/progress/619.md
@@ -1,0 +1,13 @@
+# Issue #619 — Extract shared prompt helpers into internal/agent/prompt
+
+- [x] Create internal/agent/prompt/prompt.go (WriteSystemPromptFile, IsSuspiciousOutput, MinValidOutputLen, RetryDelay)
+- [x] Create internal/agent/prompt/plan.go (PlanningPromptTemplate, PromptData, RenderPlanningPrompt, ReadPlanFile, ValidatePlan)
+- [x] Create internal/agent/prompt/review.go (ReviewPromptTemplate, ReviewSystemPrompt, ReviewPromptData, RenderReviewPrompt, ParseReviewOutput)
+- [x] Create internal/agent/prompt/context.go (gatherDirTree, gatherKeyFile, gatherGitLog, detectManifestFile, maxFileChars, truncate)
+- [x] Create internal/agent/prompt/prompt_test.go, plan_test.go, review_test.go, context_test.go
+- [x] Refactor claude package (claude.go, plan.go, review.go, execute.go, discuss.go, delete context.go) to use prompt package
+- [x] Update/move claude tests
+- [x] go build ./... passes
+- [x] go vet ./... passes
+- [x] go test ./... passes
+- [x] golangci-lint run passes
