+++ internal/agent/prompt/context.go
@@ -1,4 +1,4 @@
-package claude
+package prompt
 
 import (
 	"fmt"
